Changed Main
No sleep. Added some early structure for grabbing an input text from args to main. Added a check for if there is a useable file. Created a try-catch and created and closed a scanner for use. No catch yet. Will make a method to store rows and columns into matrix A, matrix B, and matrix C