Fixed error
An error kept arising where the program would not move forward because of the scanner closing. I was advised to allow Java to close the scanner on its own, but I feel weird about that. I am also in the middle of nowhere with no internet so I will just leave it until I can Google the issue.